Add documentation, working on tests
