+++ Sources/SwiftLocation/Async Tasks/LocationServicesEnabled.swift
@@ -4,15 +4,34 @@ import CoreLocation
 extension Tasks {
     
     public final class LocationServicesEnabled: AnyTask {
+        
+        // MARK: - Support Structures
+        
+        /// Stream produced by the task.
         public typealias Stream = AsyncStream<StreamEvent>
         
+        /// The event produced by the stream.
         public enum StreamEvent {
+            
+            /// A new change in the location services status has been detected.
             case didChangeLocationEnabled(_ enabled: Bool)
+            
+            /// Return `true` if location service is enabled.
+            var isLocationEnabled: Bool {
+                switch self {
+                case let .didChangeLocationEnabled(enabled):
+                    enabled
+                }
+            }
         }
+        
+        // MARK: - Public Properties
         
         public let uuid = UUID()
         public var stream: Stream.Continuation?
         public var cancellable: CancellableTask?
+        
+        // MARK: - Functions
         
         public func receivedLocationManagerEvent(_ event: LocationManagerBridgeEvent) {
             switch event {
