+++ Sources/SwiftLocation/Location Managers/LocationAsyncBridge.swift
@@ -1,26 +1,37 @@
 import Foundation
 import CoreLocation
 
+/// This bridge is used to link the object which manage the underlying events
+/// from `CLLocationManagerDelegate`.
 final class LocationAsyncBridge: CancellableTask {
     
-    var tasks = [AnyTask]()
-    private weak var location: SwiftLocation?
+    // MARK: - Private Properties
     
-    init(location: SwiftLocation) {
-        self.location = location
-    }
+    private var tasks = [AnyTask]()
+    weak var location: Location?
 
+    // MARK: - Internal function
+    
+    /// Add a new task to the queued operations to bridge.
+    ///
+    /// - Parameter task: task to add.
     func add(task: AnyTask) {
         task.cancellable = self
         tasks.append(task)
         task.willStart()
     }
     
+    /// Cancel the execution of a task.
+    ///
+    /// - Parameter task: task to cancel.
     func cancel(task: AnyTask) {
         cancel(taskUUID: task.uuid)
     }
     
-    func cancel(taskUUID uuid: UUID) {
+    /// Cancel the execution of a task with a given unique identifier.
+    ///
+    /// - Parameter uuid: unique identifier of the task to remove
+    private func cancel(taskUUID uuid: UUID) {
         tasks.removeAll { task in
             if task.uuid == uuid {
                 task.didCancelled()
@@ -31,19 +42,28 @@ final class LocationAsyncBridge: CancellableTask {
         }
     }
     
+    /// Cancel the task of the given class and optional validated condition.
+    ///
+    /// - Parameters:
+    ///   - type: type of `AnyTask` conform task to remove.
+    ///   - condition: optional condition to verify in order to cancel.
     func cancel(tasksTypes type: AnyTask.Type, condition: ((AnyTask) -> Bool)? = nil) {
         let typeToRemove = ObjectIdentifier(type)
         tasks.removeAll(where: {
-            let isCorrectType = $0.taskType == typeToRemove
+            let isCorrectType = ($0.taskType == typeToRemove)
+            let isConditionValid = (condition == nil ? true : condition!($0))
+            let shouldRemove = (isCorrectType && isConditionValid)
             
-            guard let condition else {
-                return isCorrectType
+            if shouldRemove {
+                $0.didCancelled()
             }
-            
-            return (isCorrectType && condition($0))
+            return shouldRemove
         })
     }
     
+    /// Dispatch the event to the tasks.
+    ///
+    /// - Parameter event: event to dispatch.
     func dispatchEvent(_ event: LocationManagerBridgeEvent) {
         for task in tasks {
             task.receivedLocationManagerEvent(event)
