+++ Sources/SwiftLocation/Support/Errors.swift
@@ -1,10 +1,21 @@
 import Foundation
 
+/// Throwable errors
 enum Errors: LocalizedError {
+    
+    /// Info.plist authorization are not correctly defined.
     case plistNotConfigured
+    
+    /// System location services are disabled by the user or not available.
     case locationServicesDisabled
+    
+    /// You must require location authorization from the user before executing the operation.
     case authorizationRequired
+    
+    /// Not authorized by the user.
     case notAuthorized
+    
+    /// Operation timeout.
     case timeout
     
     var errorDescription: String? {
